Move VPNConnectionProtocol to its own file
